# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.1.2'
  title: Time To Fish
host: localhost:3010
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /menus:
    x-swagger-router-controller: menu-items.server.controller
    get:
      operationId: getAll
      summary: 'Returns all menu items, sorted by: key, parent->child, sequence, name (alphabetical)'
      parameters:
        - name: key
          type: string
          in: query
          required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/GetMenuItemsResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      operationId: create
      summary: 'Create a new menu item. The newly created item is returned with the response'
      parameters:
        - name: menuItem
          description: 'Menu item definition'
          in: body
          required: true
          schema:
            $ref: '#/definitions/MenuItem'
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/GetMenuItemResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      operationId: deleteByKey
      summary: 'Delete a menu item with the given key. If [deleteChildren: false] is true, then all children of the deleted item are also deleted (this goes one level deep). If [recursive: false] is true, then the "one level deep" rule is ignored. Any children that remain after the delete have their parent set to the parent of the deleted item (the recursive rule applies here as well).'
      parameters:
        - name: key
          type: string
          format: hex
          in: query
          required: true
      responses:
        "204":
          description: Success
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /menus/{id}:
    x-swagger-router-controller: menu-items.server.controller
    get:
      operationId: getById
      summary: 'Returns a single menu item with the given key'
      parameters:
        - name: id
          type: string
          format: hex
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/GetMenuItemResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      operationId: deleteById
      summary: 'Delete a menu item with the given id. If [deleteChildren: false] is true, then all children of the deleted item are also deleted (this goes one level deep). If [recursive: false] is true, then the "one level deep" rule is ignored. Any children that remain after the delete have their parent set to the parent of the deleted item (the recursive rule applies here as well).'
      parameters:
        - name: id
          type: string
          format: hex
          in: path
          required: true
      responses:
        "204":
          description: Success
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /swagger:
    x-swagger-pipe: swagger_raw
# complex object schema definitions
definitions:
  MenuItem:
    required:
      - key
      - name
      - sequence
    properties:
      key:
        type: string
      name:
        type: string
      sequence:
        type: integer
        format: int32
      description:
        type: string
      parent:
        $ref: '#/definitions/MenuItem'
      state:
        type: string
      permissions:
        type: array
        items:
          type: string
          format: hex
  GetMenuItemResponse:
    properties:
      menuItem:
        $ref: '#/definitions/MenuItem'
  GetMenuItemsResponse:
    required:
      - menuItems
    properties:
      menuItems:
        type: array
        items:
          $ref: '#/definitions/MenuItem'
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string